WIP. Fix Row classes hierarchy.

+++ modules/table/src/main/java/org/apache/ignite/internal/table/TableSchemaManager.java
@@ -32,5 +32,5 @@ public interface TableSchemaManager {
      * @param ver Schema version.
      * @return Schema of given version.
      */
-    SchemaDescriptor schema(short ver);
+    SchemaDescriptor schema(int ver);
 }
